+++ src/GuestConfiguration/GuestConfiguration.Test/ScenarioTests/GetVmGuestPolicyStatusHistoryTests.ps1
@@ -18,8 +18,8 @@ Get guest configuration policy status history by Vm name scope
 #>
 function Get-AzVMGuestPolicyStatusHistory-VmNameScope
 {
-	$rgName = "vivga"
-	$vmName = "Viv1809SDDC"
+	$rgName = "aashishGoodPolicy"
+	$vmName = "aashishvm1"
 
     $historicalStatuses = Get-AzVMGuestPolicyStatusHistory -ResourceGroupName $rgName -VMName $vmName
 	Assert-NotNull $historicalStatuses
@@ -32,8 +32,8 @@ Get guest configuration policy status history by Vm name scope
 #>
 function Get-AzVMGuestPolicyStatusHistory-VmNameScope_Custom
 {
-	$rgName = "amits-test"
-	$vmName = "amits-winser6"
+	$rgName = "aashishCustomrole7ux"
+	$vmName = "aashishCustomrole7ux"
 
     $historicalStatuses = Get-AzVMGuestPolicyStatusHistory -ResourceGroupName $rgName -VMName $vmName
 	Assert-NotNull $historicalStatuses
@@ -46,9 +46,9 @@ Get guest configuration policy status history by Initiative id scope
 #>
 function Get-AzVMGuestPolicyStatusHistory-InitiativeIdScope
 {
-	$rgName = "vivga"
-	$vmName = "Viv1809SDDC"
-	$initiativeId = "/providers/Microsoft.Authorization/policySetDefinitions/25ef9b72-4af2-4501-acd1-fc814e73dde1"
+	$rgName = "aashishGoodPolicy"
+	$vmName = "aashishvm1"
+	$initiativeId = "/providers/Microsoft.Authorization/policySetDefinitions/8bc55e6b-e9d5-4266-8dac-f688d151ec9c"
 
     $historicalStatuses = Get-AzVMGuestPolicyStatusHistory -ResourceGroupName $rgName -VMName $vmName -InitiativeId $initiativeId
 	Assert-NotNull $historicalStatuses
@@ -61,9 +61,9 @@ Get guest configuration policy status history by Initiative id scope
 #>
 function Get-AzVMGuestPolicyStatusHistory-InitiativeIdScope_Custom
 {
-	$rgName = "amits-test"
-	$vmName = "amits-winser6"
-	$initiativeId = "/subscriptions/b5e4748c-f69a-467c-8749-e2f9c8cd3db0/providers/Microsoft.Authorization/policySetDefinitions/8174929b-2752-46af-9167-89d2d6780439"
+	$rgName = "aashishCustomrole7ux"
+	$vmName = "aashishCustomrole7ux"
+	$initiativeId = "/subscriptions/b5e4748c-f69a-467c-8749-e2f9c8cd3db0/providers/Microsoft.Authorization/policySetDefinitions/60062d3c-3282-4a3d-9bc4-3557dded22ca"
 
     $historicalStatuses = Get-AzVMGuestPolicyStatusHistory -ResourceGroupName $rgName -VMName $vmName -InitiativeId $initiativeId
 	Assert-NotNull $historicalStatuses
@@ -76,9 +76,9 @@ Get guest configuration policy status history by Initiative name scope
 #>
 function Get-AzVMGuestPolicyStatusHistory-InitiativeNameScope
 {
-	$rgName = "vivga"
-	$vmName = "Viv1809SDDC"
-	$initiativeName = "25ef9b72-4af2-4501-acd1-fc814e73dde1"
+	$rgName = "aashishGoodPolicy"
+	$vmName = "aashishvm1"
+	$initiativeName = "8bc55e6b-e9d5-4266-8dac-f688d151ec9c"
 
     $historicalStatuses = Get-AzVMGuestPolicyStatusHistory -ResourceGroupName $rgName -VMName $vmName -InitiativeName $initiativeName
 	Assert-NotNull $historicalStatuses
@@ -91,9 +91,9 @@ Get guest configuration policy status history by Initiative name scope
 #>
 function Get-AzVMGuestPolicyStatusHistory-InitiativeNameScope_Custom
 {
-	$rgName = "amits-test"
-	$vmName = "amits-winser6"
-	$initiativeName = "8174929b-2752-46af-9167-89d2d6780439"
+	$rgName = "aashishCustomrole7ux"
+	$vmName = "aashishCustomrole7ux"
+	$initiativeName = "60062d3c-3282-4a3d-9bc4-3557dded22ca"
 
     $historicalStatuses = Get-AzVMGuestPolicyStatusHistory -ResourceGroupName $rgName -VMName $vmName -InitiativeName $initiativeName
 	Assert-NotNull $historicalStatuses
@@ -106,8 +106,8 @@ Get guest configuration policy by vmName scope, using ShowOnlyChange switch
 #>
 function Get-AzVMGuestPolicyStatusHistory-ShowOnlyChangeSwitch-VmNameScope
 {
-	$rgName = "vivga"
-	$vmName = "Viv1809SDDC"
+	$rgName = "aashishGoodPolicy"
+	$vmName = "aashishvm1"
 
     $historicalStatuses = Get-AzVMGuestPolicyStatusHistory -ResourceGroupName $rgName -VMName $vmName -ShowOnlyChange
 	Assert-NotNull $historicalStatuses
@@ -120,8 +120,8 @@ Get guest configuration policy by vmName scope, using ShowOnlyChange switch
 #>
 function Get-AzVMGuestPolicyStatusHistory-ShowOnlyChangeSwitch-VmNameScope_Custom
 {
-	$rgName = "amits-test"
-	$vmName = "amits-winser6"
+	$rgName = "aashishCustomrole7ux"
+	$vmName = "aashishCustomrole7ux"
 
     $historicalStatuses = Get-AzVMGuestPolicyStatusHistory -ResourceGroupName $rgName -VMName $vmName -ShowOnlyChange
 	Assert-NotNull $historicalStatuses