+++ src/ResourceGraph/ResourceGraph/ChangeLog.md
@@ -19,10 +19,10 @@
 -->
 ## Upcoming Release
 * Updated package Microsoft.Azure.Management.ResourceGraph to version 2.0
+* Adding param to extend query result with subscription and tenant names
 
 ## Version 0.7.3
 * Improving logic of getting subscriptions for query to ARG
-* Adding param to extend query result with subscription and tenant names
 
 ## Version 0.7.2
 * Fix conversion to PSCustomObject[] for top-level arrays