+++ src/ResourceGraph/ResourceGraph/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated package Microsoft.Azure.Management.ResourceGraph to version 2.0
+
+## Version 0.7.3
+* Improving logic of getting subscriptions for query to ARG
 
 ## Version 0.7.2
 * Fix conversion to PSCustomObject[] for top-level arrays